# k-vet-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "k-vet API",
    "description": "Veterinary practice management — single user, self-hosted.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/services": {
      "get": {
        "tags": [
          "services"
        ],
        "operationId": "listServices",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Text filter (name, number, …).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archived",
            "in": "query",
            "description": "Include archived records (default: false).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "hidden",
            "in": "query",
            "description": "Include hidden services (the surgical GOT positions, by default).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ServiceType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "services"
        ],
        "operationId": "createService",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateService"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "An empty draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          }
        }
      }
    },
    "/api/services/{id}": {
      "get": {
        "tags": [
          "services"
        ],
        "operationId": "getService",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "services"
        ],
        "operationId": "patchService",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchService"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        }
      }
    },
    "/api/services/{id}/archive": {
      "post": {
        "tags": [
          "services"
        ],
        "operationId": "archiveService",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          }
        }
      }
    },
    "/api/services/{id}/unarchive": {
      "post": {
        "tags": [
          "services"
        ],
        "operationId": "unarchiveService",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateService": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/ServiceType"
          }
        }
      },
...
      "PatchService": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "got_number": {
            "type": [
              "string",
              "null"
            ]
          },
          "factor": {
            "type": [
              "string",
              "null"
            ]
          },
          "vat_percent": {
            "type": [
              "string",
              "null"
            ]
          },
          "net_price": {
            "type": [
              "string",
              "null"
            ]
          },
          "travel_expenses": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "hidden": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
      "Service": {
        "type": "object",
        "required": [
          "id",
          "type",
          "travel_expenses",
          "hidden",
          "archived",
          "draft",
          "missing_fields",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "$ref": "#/components/schemas/ServiceType",
            "description": "`got` (official fee schedule) or `self_defined`."
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "got_number": {
            "type": [
              "string",
              "null"
            ]
          },
          "factor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Percent; 100 is the single rate."
          },
          "vat_percent": {
            "type": [
              "string",
              "null"
            ]
          },
          "net_price": {
            "type": [
              "string",
              "null"
            ],
            "description": "The fee **without** VAT — the GOT publishes net, so that is what is stored and edited."
          },
          "gross_price": {
            "type": [
              "string",
              "null"
            ],
            "description": "Derived from `net_price` and `vat_percent` so the UI can show what the customer pays."
          },
          "travel_expenses": {
            "type": "boolean",
            "description": "Travel-expense lines are priced from the kilometres entered (FR-023)."
          },
          "hidden": {
            "type": "boolean",
            "description": "Hidden services stay out of pickers and lists by default (FR-021)."
          },
          "archived": {
            "type": "boolean"
          },
          "draft": {
            "type": "boolean"
          },
          "missing_fields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ServiceType": {
        "type": "string",
        "description": "Official fee schedule position versus a self-defined service.",
        "enum": [
          "got",
          "self_defined"
        ]
      },
...
    }
  },
...
}