# k-vet-web/src/api/generated/model/patchService.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * k-vet API
 * Veterinary practice management — single user, self-hosted.
 * OpenAPI spec version: 0.1.0
 */

export interface PatchService {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  got_number?: string | null;
  /** @nullable */
  factor?: string | null;
  /** @nullable */
  vat_percent?: string | null;
  /** @nullable */
  net_price?: string | null;
  /** @nullable */
  travel_expenses?: boolean | null;
  /** @nullable */
  hidden?: boolean | null;
}