# k-vet-web/src/api/generated/model/service.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * k-vet API
 * Veterinary practice management — single user, self-hosted.
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceType } from './serviceType';

export interface Service {
  id: number;
  /** `got` (official fee schedule) or `self_defined`. */
  type: ServiceType;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  got_number?: string | null;
  /**
     * Percent; 100 is the single rate.
     * @nullable
     */
  factor?: string | null;
  /** @nullable */
  vat_percent?: string | null;
  /**
     * The fee **without** VAT — the GOT publishes net, so that is what is stored and edited.
     * @nullable
     */
  net_price?: string | null;
  /**
     * Derived from `net_price` and `vat_percent` so the UI can show what the customer pays.
     * @nullable
     */
  gross_price?: string | null;
  /** Travel-expense lines are priced from the kilometres entered (FR-023). */
  travel_expenses: boolean;
  /** Hidden services stay out of pickers and lists by default (FR-021). */
  hidden: boolean;
  archived: boolean;
  draft: boolean;
  missing_fields: string[];
  created_at: string;
  updated_at: string;
}